requirements.txt
Flask
deep-translator
pandas
numpy
joblib
scikit-learn
catboost
lightgbm
groq
python-dotenv
